refactor: move serialport functionality into separate file
