+++ src/serverroutes.js
@@ -26,7 +26,7 @@ const { getHttpPort, getSslPort } = require('./ports')
 const express = require('express')
 const { getAISShipTypeName } = require('@signalk/signalk-schema')
 const { queryRequest } = require('./requestResponse')
-const SerialPort = require('serialport')
+const { listAllSerialPorts } = require('./serialports')
 const commandExists = require('command-exists')
 const { getAuthor, restoreModules } = require('./modules')
 const zip = require('express-easy-zip')
@@ -567,49 +567,11 @@ module.exports = function(app, saveSecurityConfig, getSecurityConfig) {
   })
 
   app.get('/serialports', (req, res, next) => {
-    Promise.all([
-      listSafeSerialPortsDevSerialById(),
-      listSafeSerialPortsDevSerialByPath(),
-      listSafeSerialPortsOpenPlotter(),
-      listSerialPorts()
-    ])
-      .then(([byId, byPath, byOpenPlotter, serialports]) =>
-        res.json({ byId, byPath, byOpenPlotter, serialports })
-      )
+    listAllSerialPorts().then(ports => {
+      res.json(ports)
+    })
       .catch(next)
   })
-
-  function listSerialPorts() {
-    return SerialPort.list().then(ports => {
-      return ports.map(port => port.comName)
-    })
-  }
-
-  function listSafeSerialPortsDevSerialById() {
-    return readdir('/dev/serial/by-id')
-      .catch(err => [])
-      .then(filenames =>
-        filenames.map(filename => `/dev/serial/by-id/${filename}`)
-      )
-  }
-
-  function listSafeSerialPortsDevSerialByPath() {
-    return readdir('/dev/serial/by-path')
-      .catch(err => [])
-      .then(filenames =>
-        filenames.map(filename => `/dev/serial/by-path/${filename}`)
-      )
-  }
-
-  function listSafeSerialPortsOpenPlotter() {
-    return readdir('/dev/')
-      .catch(err => [])
-      .then(filenames =>
-        filenames
-          .filter(filename => filename.startsWith('ttyOP_'))
-          .map(filename => `/dev/${filename}`)
-      )
-  }
 
   app.get(`${serverRoutesPrefix}/hasAnalyzer`, (req, res) => {
     commandExists('analyzer')
